refactor: dont export project from constructs [sc-0]
